requirements.txt
torch
TTS
piper-tts
transformers>=4.31.0 # Bark vereist vaak een recente versie
accelerate # Optioneel, maar kan helpen met Bark prestaties
scipy # Nodig voor Bark om WAV op te slaan
soundfile # Alternatief voor 'wave' module, soms robuuster
pydub # Voor de optionele XTTS chunk concatenatie
numpy
pygame
elevenlabs
python-dotenv
# Zorg ervoor dat je CUDA toolkit hebt geïnstalleerd als je GPU wilt gebruiken